Added the ability to play audio files on top of the slideshow.
Any audio files found in the selected file folder will be played as the slide is running. Audio files can be added and removed live just like video or image files.

If video sounds are not muted, audio files will pause while a video is playing.

Audio files respect the shuffle option and will play in order if it's false or randomly if it's true